Accounts have a kind and safe entries
Added kind so we can sort accounts by asset, liability, equity.
Seems best to hold this as a string instead of relying upon
table relations.

Added safe methods and tests to see if the new locks, and checks
work.

Updated the transaction balance check to A = L + E formula to
ensure consistency in the system
:wq